Add method to check for GCC function attributes
It's fairly common on Linux and *BSD platforms to check for these
attributes existence, so it makes sense to me to have this checking
build into meson itself. Autotools also has a builtin for handling
these, and by building them in we can short circuit cases that we know
that these don't exist (MSVC).

The Intel C Compiler is interesting, it appears that on Linux it has
support for __attribute__, but not on windows, because on Linux it tries
to behave like GCC, but on Windows it tries to behave like MSVC.

This isn't nearly all of the attributes that GCC supports, but it is all
the attributes that autotools has built in rules for.