wip: Add a module to parse automake files
Specifically, this is meant to provide the ability to parse makefiles
that consist of nothing but variables consisting of sources. The goal is
to be able to have a single Makefile.sources file that can be included
in an autotools build system, an Android.mk build system, and a meson
build system, and not have to duplicate the sources in meson.

For many projects it's possible to drop autotools immediately, for
others autotools support may be needed for some time. For core system
libraries like mesa and xorg that are used on operating systems beyond
Linux and Windows and/or are needed by LTS and EL Linux distributions
maintaining an autotools build system for some time is absolutely
necessary. For mesa, android support is also necessary, and we have no
control over what build system they use, but duplicating sources across
3 build system would be completely intractable.

While it might be possible to hack something into autotools to
understand a new format, it's not possible to do the same for
android.mk, so it makes more sense to meson to implement a format that
is already common to other build systems that projects need to support.