dependencies: Add a dependency factory
One of the things that really bugged me is the way we handle
dependencies that can be found multiple ways. We create an
ExternalDependency object, and then create a PkgConfigDependency,
ConfigToolDependency, and/or ExtraFrameworksDependency, then set the
attributes of the ExternalDependency. This is very odd and very fragile,
since any new method to any of those types has to be set in each of
these instances (and we keep adding more).

Instead, I propsoe using a factory class for these kind of dependencies.
This factory is able to create the different types of dependencies and
return the leaf class type rather than assigning it's own attributes
from the leaf class. This is a much nicer pattern, it reduces the amount
of code that needs to be duplicated and it removes the fragility of the
assignment we're doing currently.