Convert Intel Compiler to Intel Nix Compiler
ICC is a bit special, in that it tries to behave like MSVC on windows
and like GCC on Linux and MacOS, including command line options. I think
that it'll prove to be an exercise in masochism to try to share a
GNUlike abc between windows and !Windows therefore. Let's rename the
Intel stuff to IntelNix and move on.