compilers: Make CompilerArgs not crazy
Currently CompilerArgs allows passing 0, 1, or 2 parameters. In any
order. Which is crazy. We never use the 0 case, so just throw that out.
Also ensure that the first argument is a Compiler or StaticLinker, and
that is required. The second argument is now an optional list of
initial values. This makes the thing easier to reason about and use, and
makes it possible for linters to actually figure out what's going on.