replace uses of map and filter with comprehensions
map and filter are semi deprecated in python in favor of comprehensions;
in part because comprehensions are much more readable than map/filter,
but also because they are much faster. Apart from being better optimized
they avoid the need for an extra function call per iteration, especially
when than call is a lambda (all the slow).