compilers/d: Implement linker overriding for dmd
This is pretty tricky to do. DMD only supports overriding the linker via
environment variables (which we don't want to support) and via a config
file (no command line switches like gdc/ldc). To get this to work we
have to resolve the path to that file, then read the file, replace a
bunch of $PWD-esqe variables, and finally set the linker variable. This
this config can be used to override the linker with a supported linker
(optlink, link, lld-link).

Does not work yet: my method for detecting the architecture doesn't
work, it depends on information that isn't yet available.