dependencies/misc: Overhaul the pythonh3 dependency
This makes two releated changes. First it pulls in the python dependency
finder code for the sysconfig version of python from the python module,
the second is it instructs the pkg-config version of python to get all
of the installed versions of python from pkg-config, and try all of
those rather than just python3.pc, which may or may not even exist
(gentoo doesn't have one).