modules/python: Use the python_factory for getting a dependency
This means we use the same code in both the module and in dependency(),
this de-deduplicates code and lowers maintenance. This also means that
the only difference between calling `dependency('python')` and
`py_installation.dependency()` is that the py_installation will set a
default `version` argument and warn if one is provided.

Fixes #6792