modules/python: Allow using python2 or python3 for installation name
If you have a strange project that needs both versions, or you
explicitly want to use 'python3' it's confusing that the entry in the
machine file is `python` not `python2` or `python3`. This change alloiw
`python2` or `python3` to be used *if* an argument is passed to
`find_installation`, and that argument contains `python2` or `python3`
respectively.

Partial Fix: #7078