modules/python: Add a version kwarg to find_installation
This allows the program authors to have more control over their python
versions (such as if they require 3.7 and new features that have no
backports). I've also refactored the method to be smarter about finding
python in cases where modules are requested, meson now considers modules
to be part of the found condition rather than something to be decided
after finding the version of python and then failing when the modules
don't exist.

Fixes: #7078