compilers: Get rid of add_global_args
Because we now have all of the options stored in the environment (at
least the relevant ones like c_args and c_link_args), we don't need to
initialize the options before we look for the linker, and don't need
this. Instead the linker search code can just read the options from
Environment (which is where it is anyway).