modules: Add a rust module
Currently it has one function, test. This method allows defining rust
tests in a DRY way. Instead of creating two identical targets, one wit
--test added and one without, you create only the main target, then pass
that target to the special rust function which creates a new executable
using the other as a template, and adds --test, then generates the test
target as well, making for a much DRYer experience.