modules: Add a rust modlue
Like other language specific modules this module is module for holding
rust specific helpers. This commit adds a test() function, which
simplifies using rust's internal unittest mechanism.

This mechanism requires compiling a given set of rust files twice, once
as whatever you want to use them as (either a library or executable) and
a second time a test executable. Generally these are compiled with the
same arguments, except for the addition of `--test`. To simplify this
the rustmod.test() function takes the target you want as an input, and
copies the sources and arguments from the original target, then creates
a new executable with that information, plus the --test arguments, then
it creates a Test() target with the rust protocol automatically. This
makes for nice, DRY, test definitions.