typed_kwargs: check for required before checking for a valid type
Often we want to use a pattern like
```python
@typed_kwargs(
    'foo',
    KwargInfo('foo', str, required=True)
)
def foo(...): ...
```

This is nice in that we don't annotate that NoneType is allowed, even
though the default is None, since the function will never actually
receive a value of None, it will only receive a string value. Currently
this fails with an assertion error, since we check the value type first,
then check for required.

Fixes: #9698