compilers: Remove system paths from compiler search dirs
For some reason GCC sometimes lists /lib and other system locations as
part of its default search path. This shouldn't be necessary, as GCC
shouldn't be installing anything in there, and if we need it other
dependencies will add it. It's a problem because it sometimes prevents
the correct dependency from being found.