modules/codegen: Add wrapper for lex
This module is a bit of a dumping ground for code generators,
particularly ones that are important and non-trivial to wrap, either due
to multiple implementations, major command line changes, or complex
outputs (such as those that may output a directory structure).

The initially provided method is for lex. It provides a simple wrapper
that handles win_flex, reflex, flex, and generic lex.