modules/export: Add a stub for an export module
This module is meant to provide a single interface for exporting
dependencies. The goal of this module is to provide a single method for
getting pkg-config data and cps data, but possibly other formats could
be handled if necessary. This should provide a better user experience,
since new formats could be handled transparently by simply updating,
without the need to know which formats they're supporting.