compilers/cuda: Fix sanity_check when running with a relative build dir
Currently, this check adds the working directory to the paths of the
generated files it wants to use. This works when run inside mesonmain
because we always have an absolute build directory, but when put into
run_project_tests.py it doesn't work because that gives a relative build
directory.